Enlazando el Backend con  las paginas de Mis Solicitudes y Tracking en la vista Cliente

+++ Backend/b2b/src/main/java/com/nethink/b2b/dto/request/ItemCotizadoRequest.java
@@ -1,8 +1,10 @@
 package com.nethink.b2b.dto.request;
 
+import java.math.BigDecimal;
+
 
 public record ItemCotizadoRequest(
     Integer idProducto,
     Integer cantidad,
-    Double precioUnitario
+    BigDecimal precioUnitario
 ) {}
